requirements.txt
amqp==2.6.0
anaconda-client==1.2.2
appdirs==1.4.3
asgiref==3.2.7
billiard==3.6.3.0
CacheControl==0.12.6
cairocffi==1.1.0
CairoSVG==2.4.2
celery==4.4.5
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
clyent==1.2.1
colorama==0.4.3
contextlib2==0.6.0
coverage==5.2.1
cssselect2==0.3.0
defusedxml==0.6.0
diff-match-patch==20200713
distlib==0.3.0
distro==1.4.0
dj-database-url==0.5.0
Django==3.0.6
django-appconf==1.0.4
django-celery-email==3.0.0
django-crispy-forms==1.9.0
django-import-export==2.3.0
djangorestframework==3.11.1
et-xmlfile==1.0.1
flower==0.9.4
future==0.18.2
gunicorn==20.0.4
html5lib==1.0.1
humanize==0.5.1
idna==2.9
ipaddr==2.2.0
jdcal==1.4.1
kombu==4.6.10
lockfile==0.12.2
MarkupPy==1.14
msgpack==0.6.2
odfpy==1.4.1
openpyxl==3.0.4
packaging==20.3
pep517==0.8.2
Pillow==7.1.2
progress==1.5
prometheus-client==0.8.0
psycopg2==2.8.5
pycparser==2.20
Pygments==2.6.1
PyJWT==1.7.1
pyparsing==2.4.6
Pyphen==0.9.5
python-dateutil==2.8.1
python-decouple==3.3
python-dotenv==0.13.0
pytoml==0.1.21
pytz==2020.1
PyYAML==5.3.1
redis==3.5.3
reportlab==3.5.42
requests==2.23.0
retrying==1.3.3
sentry-sdk==0.17.3
six==1.14.0
sqlparse==0.3.1
tablib==2.0.0
tinycss2==1.0.2
toml==0.10.1
tornado==6.0.4
twilio==6.45.1
urllib3==1.25.9
uWSGI==2.0.19.1
vine==1.3.0
WeasyPrint==51
webencodings==0.5.1
whitenoise==5.1.0
xlrd==1.2.0
xlwt==1.3.0
